Propagate HIP GPU arches throughout DGL builds (#37)
This attempts to get all of the many ways that a CMake build with HIP
can specify the GPU architectures to propagate across the various
sub-builds DGL creates. I'm honestly not sure how much of this is
necessary. I *think* that `CMAKE_HIP_ARCHITECTURES` is supposed to be
sufficient and the new blessed way
(https://rocm.docs.amd.com/en/latest/conceptual/cmake-packages.html),
but HIP CMake complains anyways
(https://github.com/ROCm/ROCm/issues/4296). To be thorough I have just
set all of these to the same thing everywhere, propagating from the
parent build.

The HIP language mode requires CMake >=3.21, so we also check for that
if ROCm is enabled.